backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
tortoise-orm>=0.20.0
aerich>=0.7.2
aiomysql>=0.2.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Auth
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# Crypto
pycryptodome>=3.20.0

# HTTP client
httpx>=0.26.0

# Config & Logging
python-dotenv>=1.0.0
loguru>=0.7.2

# Data processing
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
# pandas_ta>=0.3.14b0  # Requires Python 3.12+
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Image processing
Pillow>=9.0.0

# AI
openai>=1.0.0

# YAML parsing for Skills
pyyaml>=6.0
